+++ kong/tools/queue_schema.lua
@@ -6,7 +6,9 @@ return Schema.define {
   fields = {
     { name = {
       type = "string",
-      -- description = "name of the queue",
+      -- description = "name of the queue, unique across one workspace.",
+      -- If two plugin instances use the same queue name, they will
+      -- share one queue and their queue related configuration must match.
     } },
     { batch_max_size = {
       type = "number",
@@ -17,13 +19,15 @@ return Schema.define {
       type = "number",
       default = 1,
       -- description = "maximum number of (fractional) seconds to elapse after the first entry was queued before the queue starts calling the handler",
+      -- This parameter has no effect if `batch_max_size` is 1, as queued entries will be sent
+      -- immediately in that case.
     } },
-    { capacity = {
+    { max_entries = {
       type = "number",
       default = 10000,
       -- description = "maximum number of entries that can be waiting on the queue",
     } },
-    { string_capacity = {
+    { max_bytes = {
       type = "number",
       default = nil,
       -- description = "maximum number of bytes that can be waiting on a queue, requires string content",